+++ app/src/main/java/com/gemini/energy/data/local/model/ComputableLocalModel.kt
@@ -3,9 +3,9 @@ package com.gemini.energy.data.local.model
 data class ComputableLocalModel(
         var auditId: Long,
         var auditName: String,
-        var zoneId: Int,
+        var zoneId: Long,
         var zoneName: String,
-        var auditScopeId: Int,
+        var auditScopeId: Long,
         var auditScopeName: String,
         var auditScopeType: String,
         var auditScopeSubType: String
