+++ app/src/main/java/com/gemini/energy/data/repository/TypeRepository.kt
@@ -12,8 +12,8 @@ class TypeRepository(
         private val typeRemoteDataSource: TypeRemoteDataSource,
         private val auditScopeMapper: TypeMapper) {
 
-    fun get(id: Int): Observable<TypeLocalModel> = typeLocalDataSource.get(id)
-    fun getAllTypeByZone(id: Int, type: String): Observable<List<TypeLocalModel>> =
+    fun get(id: Long): Observable<TypeLocalModel> = typeLocalDataSource.get(id)
+    fun getAllTypeByZone(id: Long, type: String): Observable<List<TypeLocalModel>> =
             typeLocalDataSource.getAllTypeByZone(id, type)
     fun getAllTypeByAudit(id: Long): Observable<List<TypeLocalModel>> =
             typeLocalDataSource.getAllTypeByAudit(id)
@@ -22,8 +22,8 @@ class TypeRepository(
             typeLocalDataSource.save(auditScopeMapper.toLocal(auditScope))
 
     fun update(type: Type): Observable<Unit> = typeLocalDataSource.update(auditScopeMapper.toLocal(type))
-    fun delete(id: Int): Observable<Unit> = typeLocalDataSource.delete(id)
-    fun deleteByZoneId(id: Int): Observable<Unit> = typeLocalDataSource.deleteByZoneId(id)
+    fun delete(id: Long): Observable<Unit> = typeLocalDataSource.delete(id)
+    fun deleteByZoneId(id: Long): Observable<Unit> = typeLocalDataSource.deleteByZoneId(id)
     fun deleteByAuditId(id: Long): Observable<Unit> = typeLocalDataSource.deleteByAuditId(id)
 
 }
