+++ app/src/main/java/com/gemini/energy/data/repository/ZoneRepository.kt
@@ -12,7 +12,7 @@ class ZoneRepository(
         private val zoneRemoteDataSource: ZoneRemoteDataSource,
         private val zoneMapper: ZoneMapper) {
 
-    fun get(id: Int): Observable<ZoneLocalModel> = zoneLocalDataSource.get(id)
+    fun get(id: Long): Observable<ZoneLocalModel> = zoneLocalDataSource.get(id)
     fun getAllByAudit(id: Long): Observable<List<ZoneLocalModel>> {
         return zoneLocalDataSource.getAllByAudit(id)
     }
@@ -20,7 +20,7 @@ class ZoneRepository(
     fun save(zone: Zone): Observable<Unit> = zoneLocalDataSource.save(zoneMapper.toLocal(zone))
     fun update(zone: Zone): Observable<Unit> = zoneLocalDataSource.update(zoneMapper.toLocal(zone))
 
-    fun delete(id: Int): Observable<Unit> = zoneLocalDataSource.delete(id)
+    fun delete(id: Long): Observable<Unit> = zoneLocalDataSource.delete(id)
     fun deleteByAuditId(id: Long): Observable<Unit> = zoneLocalDataSource.deleteByAuditId(id)
 
 }
